requirements.txt
aiofiles==0.5.0
aiohttp==3.6.2
ase==3.20.1
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
dgl==0.5.2
diffbot-enhance-client==1.0.0
filelock==3.0.12
future==0.18.2
idna==2.10
joblib==0.17.0
kiwisolver==1.3.1
matplotlib==3.3.2
multidict==4.7.6
networkx==2.5
numpy==1.19.2
packaging==20.4
pandas==1.1.2
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.28
requests==2.24.0
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.2
sentencepiece==0.1.94
six==1.15.0
threadpoolctl==2.1.0
tokenizers==0.8.1rc2
torch==1.6.0
tqdm==4.50.0
transformers==3.1.0
typing-extensions==3.7.4.3
urllib3==1.25.11
yarl==1.6.2
